requirements.txt
#Alle benötigten Python packages
numpy
PIL
torch
torchvision
matplotlib.pyplot
pickle
tqdm

